+++ benches/benches/block_target_gas.rs
@@ -244,7 +244,7 @@ fn service_with_many_contracts(
     state_size: u64,
     contract_ids: Vec<ContractId>,
 ) -> (FuelService, tokio::runtime::Runtime) {
-    use fuel_core_database::vm_database::IncreaseStorageKey;
+    use fuel_core_database::vm_storage::IncreaseStorageKey;
     let rt = tokio::runtime::Builder::new_current_thread()
         .enable_all()
         .build()
