+++ benches/benches/state.rs
@@ -6,7 +6,7 @@ use criterion::{
     Criterion,
 };
 use fuel_core::database::Database;
-use fuel_core_database::vm_database::VmDatabase;
+use fuel_core_storage::vm_storage::VmStorage;
 use fuel_core_types::{
     blockchain::header::GeneratedConsensusFields,
     fuel_tx::Bytes32,
@@ -51,7 +51,7 @@ fn insert_state_single_contract_database(c: &mut Criterion) {
 
     let mut bench_state = |group: &mut BenchmarkGroup<WallTime>, name: &str, n: usize| {
         group.bench_function(name, |b| {
-            let mut db = VmDatabase::default();
+            let mut db = VmStorage::default();
             let contract: ContractId = rng.gen();
             setup(db.database_mut(), &contract, n);
             let outer = db.database_mut().transaction();
@@ -59,7 +59,7 @@ fn insert_state_single_contract_database(c: &mut Criterion) {
                 let mut elapsed_time = Duration::default();
                 for _ in 0..iters {
                     let mut inner = outer.transaction();
-                    let mut inner_db = VmDatabase::new::<GeneratedConsensusFields>(
+                    let mut inner_db = VmStorage::new::<GeneratedConsensusFields>(
                         inner.as_mut().clone(),
                         &Default::default(),
                         Default::default(),
@@ -112,7 +112,7 @@ fn insert_state_single_contract_transaction(c: &mut Criterion) {
 
     let mut bench_state = |group: &mut BenchmarkGroup<WallTime>, name: &str, n: usize| {
         group.bench_function(name, |b| {
-            let mut db = VmDatabase::<Database>::default();
+            let mut db = VmStorage::<Database>::default();
             let contract: ContractId = rng.gen();
             let mut outer = db.database_mut().transaction();
             setup(outer.as_mut(), &contract, n);
@@ -120,7 +120,7 @@ fn insert_state_single_contract_transaction(c: &mut Criterion) {
                 let mut elapsed_time = Duration::default();
                 for _ in 0..iters {
                     let mut inner = outer.transaction();
-                    let mut inner_db = VmDatabase::new::<GeneratedConsensusFields>(
+                    let mut inner_db = VmStorage::new::<GeneratedConsensusFields>(
                         inner.as_mut().clone(),
                         &Default::default(),
                         Default::default(),
@@ -173,7 +173,7 @@ fn insert_state_multiple_contracts_database(c: &mut Criterion) {
 
     let mut bench_state = |group: &mut BenchmarkGroup<WallTime>, name: &str, n: usize| {
         group.bench_function(name, |b| {
-            let mut db = VmDatabase::<Database>::default();
+            let mut db = VmStorage::<Database>::default();
             for _ in 0..n {
                 let contract: ContractId = rng.gen();
                 setup(db.database_mut(), &contract, 1);
@@ -184,7 +184,7 @@ fn insert_state_multiple_contracts_database(c: &mut Criterion) {
                 let contract: ContractId = rng.gen();
                 for _ in 0..iters {
                     let mut inner = outer.transaction();
-                    let mut inner_db = VmDatabase::new::<GeneratedConsensusFields>(
+                    let mut inner_db = VmStorage::new::<GeneratedConsensusFields>(
                         inner.as_mut().clone(),
                         &Default::default(),
                         Default::default(),
@@ -237,7 +237,7 @@ fn insert_state_multiple_contracts_transaction(c: &mut Criterion) {
 
     let mut bench_state = |group: &mut BenchmarkGroup<WallTime>, name: &str, n: usize| {
         group.bench_function(name, |b| {
-            let mut db = VmDatabase::<Database>::default();
+            let mut db = VmStorage::<Database>::default();
             let mut outer = db.database_mut().transaction();
             for _ in 0..n {
                 let contract: ContractId = rng.gen();
@@ -248,7 +248,7 @@ fn insert_state_multiple_contracts_transaction(c: &mut Criterion) {
                 let contract: ContractId = rng.gen();
                 for _ in 0..iters {
                     let mut inner = outer.transaction();
-                    let mut inner_db = VmDatabase::new::<GeneratedConsensusFields>(
+                    let mut inner_db = VmStorage::new::<GeneratedConsensusFields>(
                         inner.as_mut().clone(),
                         &Default::default(),
                         Default::default(),
