+++ crates/database/src/lib.rs
@@ -6,12 +6,10 @@
 
 #![deny(clippy::arithmetic_side_effects)]
 #![deny(clippy::cast_possible_truncation)]
-// #![deny(missing_docs)]
+#![deny(missing_docs)]
 #![deny(unused_crate_dependencies)]
 #![deny(warnings)]
-#![allow(unused_variables)]
-
-pub mod vm_database;
+#![deny(unused_variables)]
 
 use fuel_core_storage::Error as StorageError;
 use fuel_core_types::services::executor::Error as ExecutorError;
