+++ crates/fuel-core/src/database/block.rs
@@ -12,7 +12,6 @@ use crate::database::{
     Result as DatabaseResult,
 };
 use fuel_core_storage::{
-    database::FuelBlockTrait,
     iter::IterDirection,
     not_found,
     tables::{
@@ -120,22 +119,6 @@ impl StorageMutate<FuelBlocks> for Database {
         }
 
         Ok(prev)
-    }
-}
-
-impl FuelBlockTrait for Database {
-    type Error = StorageError;
-
-    fn latest_height(&self) -> Result<BlockHeight, Self::Error> {
-        self.latest_height()
-    }
-
-    fn block_time(&self, height: &BlockHeight) -> Result<Tai64, Self::Error> {
-        self.block_time(height)
-    }
-
-    fn get_block_id(&self, height: &BlockHeight) -> Result<Option<BlockId>, Self::Error> {
-        self.get_block_id(height)
     }
 }
