+++ crates/fuel-core/src/database/state.rs
@@ -43,7 +43,12 @@ use std::{
     ops::Deref,
 };
 
-use fuel_core_storage::database::FuelStateTrait;
+use fuel_core_storage::vm_storage::VmStorageRequirements;
+use fuel_core_types::{
+    blockchain::primitives::BlockId,
+    fuel_types::BlockHeight,
+    tai64::Tai64,
+};
 
 impl StorageInspect<ContractsState> for Database {
     type Error = StorageError;
@@ -155,8 +160,16 @@ impl MerkleRootStorage<ContractId, ContractsState> for Database {
     }
 }
 
-impl FuelStateTrait for Database {
+impl VmStorageRequirements for Database {
     type Error = StorageError;
+
+    fn block_time(&self, height: &BlockHeight) -> Result<Tai64, Self::Error> {
+        self.block_time(height)
+    }
+
+    fn get_block_id(&self, height: &BlockHeight) -> Result<Option<BlockId>, Self::Error> {
+        self.get_block_id(height)
+    }
 
     fn init_contract_state<S: Iterator<Item = (Bytes32, Bytes32)>>(
         &mut self,
