+++ crates/services/executor/Cargo.toml
@@ -12,18 +12,14 @@ description = "Fuel Block Executor"
 [dependencies]
 anyhow = { workspace = true }
 fuel-core-chain-config = { workspace = true }
-fuel-core-database = { workspace = true }
 fuel-core-storage = { workspace = true, features = ["test-helpers"] }
 fuel-core-types = { workspace = true, features = ["test-helpers"] }
 hex = { version = "0.4", features = ["serde"] }
-itertools = { workspace = true }
 parking_lot = { workspace = true }
-rand = { workspace = true }
 tracing = { workspace = true }
 
 [dev-dependencies]
 fuel-core-trace = { path = "../../trace" }
-mockall = { workspace = true }
 
 [features]
 test-helpers = [
