+++ crates/services/executor/src/lib.rs
@@ -1,6 +1,6 @@
 #![deny(clippy::arithmetic_side_effects)]
 #![deny(clippy::cast_possible_truncation)]
-// #![deny(unused_crate_dependencies)]
+#![deny(unused_crate_dependencies)]
 #![deny(warnings)]
 
 mod config;
