+++ crates/services/executor/src/refs.rs
@@ -1,9 +1,6 @@
 mod contract;
-mod database;
 
 pub use contract::{
     ContractRef,
     ContractStorageTrait,
 };
-
-pub use database::ExecutorDatabaseTrait;
